+++ src/scope/graph/scope-graph.js
@@ -1,11 +1,14 @@
 /** @flow */
 import GraphLib from 'graphlib';
-import { BitId } from '../../bit-id';
+import { BitId, BitIds } from '../../bit-id';
 import type { ModelComponent, Version } from '../models';
 import { VERSION_DELIMITER } from '../../constants';
 import Scope from '../scope';
 import { DEPENDENCIES_TYPES, DEPENDENCIES_TYPES_UI_MAP } from '../../consumer/component/dependencies/dependencies';
 import Component from '../../consumer/component/consumer-component';
+import { getLatestVersionNumber } from '../../utils';
+import Consumer from '../../consumer/consumer';
+import ComponentsList from '../../consumer/component/components-list';
 
 const Graph = GraphLib.Graph;
 
@@ -18,9 +21,14 @@ export type DependenciesInfo = {
 
 export default class DependencyGraph {
   graph: Graph;
+  scopeName: string;
 
   constructor(graph: Object) {
     this.graph = graph;
+  }
+
+  setScopeName(scopeName: string) {
+    this.scopeName = scopeName;
   }
 
   static async loadAllVersions(scope: Scope): Promise<DependencyGraph> {
@@ -29,7 +37,7 @@ export default class DependencyGraph {
   }
 
   static async loadLatest(scope: Scope): Promise<DependencyGraph> {
-    const graph = await DependencyGraph.buildGraphOfLatestFromScope(scope);
+    const graph = await DependencyGraph.buildGraphFromScope(scope);
     return new DependencyGraph(graph);
   }
 
@@ -45,6 +53,9 @@ export default class DependencyGraph {
     return new DependencyGraph(graph);
   }
 
+  /**
+   * @todo: refactor this to work with the newer method `buildGraphFromScope`.
+   */
   static async buildGraphWithAllVersions(scope: Scope): Graph {
     const graph = new Graph({ compound: true });
     const depObj: { [id: string]: Version } = {};
@@ -77,45 +88,60 @@ export default class DependencyGraph {
     return graph;
   }
 
-  static async buildGraphOfLatestFromScope(scope: Scope): Promise<Graph> {
+  static async buildGraphFromScope(scope: Scope): Promise<Graph> {
     const graph = new Graph();
     const allModelComponents: ModelComponent[] = await scope.list();
     const buildGraphP = allModelComponents.map(async (modelComponent) => {
-      // $FlowFixMe
-      const latestVersion = await modelComponent.loadVersion(modelComponent.latest(), scope.objects);
-      const id = modelComponent.toBitIdWithLatestVersion();
-      const idStr = id.toString();
-      // save the full BitId of a string id to be able to retrieve it later with no confusion
-      if (!graph.hasNode(idStr)) graph.setNode(idStr, id);
-      DEPENDENCIES_TYPES.forEach((depType) => {
-        latestVersion[depType].get().forEach((dependency) => {
-          const depIdStr = dependency.id.toString();
-          if (!graph.hasNode(depIdStr)) graph.setNode(depIdStr, dependency.id);
-          graph.setEdge(idStr, depIdStr, depType);
-        });
+      const buildVersionP = modelComponent.listVersions().map(async (versionNum) => {
+        // $FlowFixMe
+        const version = await modelComponent.loadVersion(versionNum, scope.objects);
+        const id = modelComponent.toBitId().changeVersion(versionNum);
+        this._addDependenciesToGraph(id, graph, version);
       });
+      await Promise.all(buildVersionP);
     });
     await Promise.all(buildGraphP);
     return graph;
   }
 
-  static buildGraphFromComponents(components: Component[]): Graph {
+  static async buildGraphFromWorkspace(consumer: Consumer): Promise<Graph> {
+    const componentsList = new ComponentsList(consumer);
+    const workspaceComponents: Component[] = await componentsList.getFromFileSystem();
     const graph = new Graph();
-    components.forEach((component: Component) => {
-      const id = component.id;
-      const idStr = component.id.toString();
-      // save the full BitId of a string id to be able to retrieve it later with no confusion
-      if (!graph.hasNode(idStr)) graph.setNode(idStr, id);
-      DEPENDENCIES_TYPES.forEach((depType) => {
+    const allModelComponents: ModelComponent[] = await consumer.scope.list();
+    const buildGraphP = allModelComponents.map(async (modelComponent) => {
+      const buildVersionP = modelComponent.listVersions().map(async (versionNum) => {
         // $FlowFixMe
-        component[depType].get().forEach((dependency) => {
-          const depIdStr = dependency.id.toString();
-          if (!graph.hasNode(depIdStr)) graph.setNode(depIdStr, dependency.id);
-          graph.setEdge(idStr, depIdStr, depType);
-        });
+        const id = modelComponent.toBitId().changeVersion(versionNum);
+        const componentFromWorkspace = workspaceComponents.find(comp => comp.id.isEqual(id));
+        // if the same component exists in the workspace, use it as it might be modified
+        const version =
+          componentFromWorkspace || (await modelComponent.loadVersion(versionNum, consumer.scope.objects));
+        this._addDependenciesToGraph(id, graph, version);
       });
+      await Promise.all(buildVersionP);
+    });
+    await Promise.all(buildGraphP);
+
+    workspaceComponents.forEach((component: Component) => {
+      const id = component.id;
+      this._addDependenciesToGraph(id, graph, component);
     });
     return graph;
+  }
+
+  static _addDependenciesToGraph(id: BitId, graph: Graph, component: Version | Component): void {
+    const idStr = id.toString();
+    // save the full BitId of a string id to be able to retrieve it later with no confusion
+    if (!graph.hasNode(idStr)) graph.setNode(idStr, id);
+    DEPENDENCIES_TYPES.forEach((depType) => {
+      // $FlowFixMe
+      component[depType].get().forEach((dependency) => {
+        const depIdStr = dependency.id.toString();
+        if (!graph.hasNode(depIdStr)) graph.setNode(depIdStr, dependency.id);
+        graph.setEdge(idStr, depIdStr, depType);
+      });
+    });
   }
 
   /**
@@ -132,7 +158,8 @@ export default class DependencyGraph {
   }
 
   getDependenciesInfo(id: BitId): DependenciesInfo[] {
-    const dijkstraResults = GraphLib.alg.dijkstra(this.graph, id.toString());
+    const idWithVersion = this._getIdWithLatestVersion(id);
+    const dijkstraResults = GraphLib.alg.dijkstra(this.graph, idWithVersion.toString());
     const dependencies: DependenciesInfo[] = [];
     Object.keys(dijkstraResults).forEach((idStr) => {
       const distance = dijkstraResults[idStr].distance;
@@ -154,8 +181,9 @@ export default class DependencyGraph {
   }
 
   getDependentsInfo(id: BitId): DependenciesInfo[] {
+    const idWithVersion = this._getIdWithLatestVersion(id);
     const edgeFunc = v => this.graph.inEdges(v);
-    const dijkstraResults = GraphLib.alg.dijkstra(this.graph, id.toString(), undefined, edgeFunc);
+    const dijkstraResults = GraphLib.alg.dijkstra(this.graph, idWithVersion.toString(), undefined, edgeFunc);
     const dependents: DependenciesInfo[] = [];
     Object.keys(dijkstraResults).forEach((idStr) => {
       const distance = dijkstraResults[idStr].distance;
@@ -174,6 +202,19 @@ export default class DependencyGraph {
     });
     dependents.sort((a, b) => a.depth - b.depth);
     return dependents;
+  }
+
+  _getIdWithLatestVersion(id: BitId): BitId {
+    if (id.hasVersion()) {
+      return id;
+    }
+    const nodes = this.graph.nodes();
+    const ids = nodes.filter(n => n.startsWith(id.toString()));
+    if (!ids.length) {
+      throw new Error(`failed finding ${id.toString()} in the graph`);
+    }
+    const bitIds = ids.map(idStr => this.graph.node(idStr));
+    return getLatestVersionNumber(BitIds.fromArray(bitIds), id);
   }
 
   getComponent(id: BitId): ModelComponent {